requirements.txt
asgiref==3.7.2
attrs==23.2.0
cffi==1.16.0
cryptography==42.0.5
dj-database-url==2.1.0
Django==5.0.3
django-ckeditor-5==0.2.12
django-js-asset==2.2.0
django-mptt==0.16.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
packaging==24.0
pillow==10.2.0
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
rpds-py==0.18.0
sqlparse==0.4.4
swagger-spec-validator==3.0.3
typing_extensions==4.10.0
uritemplate==4.1.1
